refactor(coco): move analysis to a module
