+++ src/app/analysis/architecture_analysis.rs
