+++ src/app/analysis/git_analysis/format_branch.rs
