+++ src/app/mod.rs
@@ -1,6 +1,3 @@
-pub mod architecture_analysis;
-pub mod cloc_analysis;
+pub mod analysis;
 pub mod cmd;
-pub mod framework_analysis;
-pub mod git_analysis;
 pub mod visual;
